+++ packages/model/src/standard/elements/color-control.element.ts
@@ -169,7 +169,7 @@ export const ColorControl = Cluster(
 
     Attribute(
         {
-            name: "ColorLoopActive", id: 0x4002, type: "enum16", access: "R V", conformance: "CL",
+            name: "ColorLoopActive", id: 0x4002, type: "enum8", access: "R V", conformance: "CL",
             constraint: "max 1", quality: "N S"
         },
         Field({ name: "Inactive", id: 0x0 }),