+++ packages/node/src/behaviors/on-off/OnOffServer.ts
@@ -5,10 +5,12 @@
  */
 
 import { GeneralDiagnosticsBehavior } from "#behaviors/general-diagnostics";
+import { ScenesManagementServer } from "#behaviors/scenes-management";
 import { GeneralDiagnostics } from "#clusters/general-diagnostics";
 import { OnOff } from "#clusters/on-off";
 import { MaybePromise, Millis, Time, Timer } from "#general";
 import { ServerNode } from "#node/ServerNode.js";
+import { hasRemoteActor, Val } from "#protocol";
 import { OnOffBehavior } from "./OnOffBehavior.js";
 
 const OnOffLogicBase = OnOffBehavior.with(OnOff.Feature.Lighting);
@@ -41,6 +43,11 @@ export class OnOffBaseServer extends OnOffLogicBase {
                 }
             }
         }
+
+        if (this.agent.has(ScenesManagementServer)) {
+            this.agent.get(ScenesManagementServer).implementScenes(this, this.#applySceneValues);
+            this.reactTo(this.events.onOff$Changed, this.#clearDelayedSceneApplyData);
+        }
     }
 
     override async [Symbol.asyncDispose]() {
@@ -50,6 +57,9 @@ export class OnOffBaseServer extends OnOffLogicBase {
     }
 
     override on(): MaybePromise {
+        if (!this.state.onOff) {
+            this.#invalidateScenes();
+        }
         this.state.onOff = true;
         if (this.features.lighting) {
             this.state.globalSceneControl = true;
@@ -62,7 +72,15 @@ export class OnOffBaseServer extends OnOffLogicBase {
         }
     }
 
+    /** Invalidate all stored scenes manually for this endpoint in the Scenesmanagement cluster because SDK behavior. */
+    #invalidateScenes() {
+        this.agent.maybeGet(ScenesManagementServer)?.makeAllFabricSceneInfoEntriesInvalid();
+    }
+
     override off(): MaybePromise {
+        if (this.state.onOff) {
+            this.#invalidateScenes();
+        }
         this.state.onOff = false;
         if (this.features.lighting) {
             if (this.timedOnTimer.isRunning) {
@@ -91,30 +109,40 @@ export class OnOffBaseServer extends OnOffLogicBase {
     /**
      * Default implementation notes:
      * * This implementation ignores the effect and just calls off().
-     * * Global Scene Control is not supported yet.
      */
     override offWithEffect(): MaybePromise {
         if (this.state.globalSceneControl) {
-            // TODO Store state in global scene
+            if (hasRemoteActor(this.context) && this.context.session.fabric !== undefined) {
+                this.endpoint
+                    .agentFor(this.context)
+                    .maybeGet(ScenesManagementServer)
+                    ?.storeGlobalScene(this.context.session.fabric.fabricIndex);
+            }
             this.state.globalSceneControl = false;
         }
         return this.off();
     }
 
-    /**
-     * Default implementation notes:
-     * * Global Scene Control is not supported yet, so the device is just turned on.
-     */
-    override onWithRecallGlobalScene(): MaybePromise {
+    override async onWithRecallGlobalScene() {
         if (this.state.globalSceneControl) {
             return;
         }
-        // TODO Recall global scene to set onOff accordingly
+        if (
+            hasRemoteActor(this.context) &&
+            this.context.session.fabric !== undefined &&
+            this.agent.has(ScenesManagementServer)
+        ) {
+            await this.endpoint
+                .agentFor(this.context)
+                .maybeGet(ScenesManagementServer)
+                ?.recallGlobalScene(this.context.session.fabric.fabricIndex);
+        }
+
         this.state.globalSceneControl = true;
         if (this.state.onTime === 0) {
             this.state.offWaitTime = 0;
         }
-        return this.on();
+        await this.on();
     }
 
     /**
@@ -177,6 +205,45 @@ export class OnOffBaseServer extends OnOffLogicBase {
         return timer;
     }
 
+    /** Apply Scene values when requested from ScenesManagement cluster */
+    #applySceneValues(values: Val.Struct, transitionTime: number): MaybePromise {
+        this.#clearDelayedSceneApplyData();
+
+        const onOff = values.onOff;
+        // If no number (including null) or outside our range, ignore
+        if (typeof onOff !== "boolean" || this.state.onOff === onOff) {
+            return;
+        }
+
+        if (transitionTime === 0) {
+            this.state.onOff = onOff;
+        } else {
+            this.internal.applyScenePendingOnOff = onOff;
+            this.internal.applySceneDelayTimer = Time.getPeriodicTimer(
+                "delayed scene apply",
+                Millis(transitionTime),
+                this.callback(this.#applyDelayedSceneOnOffValue),
+            ).start();
+        }
+    }
+
+    #clearDelayedSceneApplyData() {
+        if (this.internal.applySceneDelayTimer?.isRunning) {
+            this.internal.applySceneDelayTimer.stop();
+        }
+        this.internal.applySceneDelayTimer = undefined;
+        this.internal.applyScenePendingOnOff = undefined;
+    }
+
+    #applyDelayedSceneOnOffValue() {
+        const onOff = this.internal.applyScenePendingOnOff;
+        this.#clearDelayedSceneApplyData();
+        if (onOff === undefined) {
+            return;
+        }
+        this.state.onOff = onOff;
+    }
+
     #delayedOffTick() {
         let time = (this.state.offWaitTime ?? 0) - 1;
         if (time <= 0) {
@@ -198,6 +265,8 @@ export namespace OnOffBaseServer {
     export class Internal {
         timedOnTimer?: Timer;
         delayedOffTimer?: Timer;
+        applySceneDelayTimer?: Timer;
+        applyScenePendingOnOff?: boolean;
     }
 }