+++ support/chip-testing/test/app-fast/S.test.ts
@@ -4,6 +4,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-describe("LVL", () => {
-    chip("LVL/*");
+describe("S", () => {
+    chip("S/*");
 });